Refactor command to separate module
