+++ src/commands/mod.rs
@@ -1,0 +1,9 @@
+//! Kwaak uses a command pattern to handle the backend asynchroniously.
+mod command;
+mod handler;
+mod responder;
+mod running_agent;
+
+pub use command::Command;
+pub use handler::CommandHandler;
+pub use responder::{CommandResponder, CommandResponse};
